review(discount): dates util
