+++ src/components/DiscountForm/DiscountForm.tsx
@@ -6,7 +6,7 @@ import Big from "big.js";
 import { cn } from "@/lib/utils";
 import { useLanguage } from "@/context/language/LanguageContext";
 import { parseFloatSafe, parseIntSafe } from "@/utils/numbers";
-import { formatDate } from "@/utils";
+import { formatDate } from "@/utils/dates";
 import { Button } from "../ui/button";
 import { FormattedCurrency } from "../FormattedCurrency";
 import { daysBetween, Act360 } from "./discount-util";
